+++ packages/agent/src/types/key-manager.ts
@@ -2,12 +2,13 @@ import type { Cipher, Jwk, KeyIdentifier, KeyWrapper, KmsExportKeyParams, KmsImp
 
 import type { Web5PlatformAgent } from './agent.js';
 import type { KeyManager } from '../prototyping/crypto/types/key-manager.js';
-import type { KeyExporter, KeyImporter } from '../prototyping/crypto/types/key-io.js';
-import type { KmsCipherParams, KmsUnwrapKeyParams, KmsWrapKeyParams } from '../prototyping/crypto/types/params-kms.js';
+import type { KeyDeleter, KeyExporter, KeyImporter } from '../prototyping/crypto/types/key-io.js';
+import type { KmsCipherParams, KmsDeleteKeyParams, KmsUnwrapKeyParams, KmsWrapKeyParams } from '../prototyping/crypto/types/params-kms.js';
 export interface AgentKeyManager extends KeyManager,
   Cipher<KmsCipherParams, KmsCipherParams>,
   KeyImporter<KmsImportKeyParams, KeyIdentifier>,
   KeyExporter<KmsExportKeyParams, Jwk>,
+  KeyDeleter<KmsDeleteKeyParams>,
   KeyWrapper<KmsWrapKeyParams, KmsUnwrapKeyParams> {
 
   agent: Web5PlatformAgent;